Fixed userRating field on Movie entity, now it works; Changed names; Implemented sorting movie by Year and Rating; Working Movie update using view, many to many relation with Genres;

+++ MoviesAPI/Models/MovieResponse.cs
@@ -11,7 +11,7 @@ namespace MoviesAPI.Models
 
         public int Year { get; set; }
 
-        public int? UsersRating { get; set; }
+        public double? UsersRating { get; set; }
 
         public Director Director { get; set; }
