+++ MoviesAPI/SqlQuery/GetDataForMovie.cs
@@ -10,7 +10,7 @@ namespace MoviesAPI.SqlQuery
         public int Id { get; private set; }
         public string Title { get; private set; }
         public int Year { get; private set; }
-        public int? UsersRating { get; private set; }
+        public double? UsersRating { get; private set; }
         public string Name { get; private set; }
         public string Surname { get; private set; }
         public string GenreName { get; private set; }
